:recycle: refactor: changing the structure
