+++ src/__init__.py
@@ -1,0 +1,4 @@
+def run():
+    from .core.dvr_viewer import DVRViewer
+    app = DVRViewer()
+    app.start()
