+++ src/core/automation.py
